# apiserver/swagger/console_swagger.yaml
...
swagger: "2.0"
info:
  description: "The API for the EatBacon IOT project"
  version: "1.0.1"
  title: "home-iot-api"
host: "localhost:4145"
basePath: "/iot/api/v1/console"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /tenants:
    post:
      tags:
      - "Tenant"
      operationId: "registerTenant"
      parameters:
      - in: "body"
        name: "tenant"
        required: false
        schema:
          $ref: "#/definitions/tenantAddRequest"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/tenantResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
    patch:
      tags:
      - "Tenant"
      operationId: "updateTenant"
      parameters:
      - in: "body"
        name: "tenants"
        required: false
        schema:
          $ref: "#/definitions/tenantAddRequest"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/tenantResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Tenant"
  /tenants/login:
    post:
      tags:
      - "Tenant"
      operationId: "loginTenant"
      parameters:
      - in: "body"
        name: "tenants"
        required: false
        schema:
          $ref: "#/definitions/tenantAddRequest"
      responses:
        200:
          description: "successfully login tenant"
        404:
          description: "The IoT hub instance or a device identity not exist"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Tenant"
  /tenants/logout:
    post:
      tags:
      - "Tenant"
      operationId: "logoutTenant"
      parameters:
      - in: "body"
        name: "tenants"
        required: false
        schema:
          $ref: "#/definitions/tenantAddRequest"
      responses:
        200:
          description: "successfully registered tenant"
      x-swagger-router-controller: "Tenant"
  /tenants/{TenantId}:
    get:
      tags:
      - "Tenant"
      operationId: "getTenant"
      parameters:
      - name: "TenantId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
    delete:
      tags:
      - "Tenant"
      operationId: "deleteTenant"
      parameters:
      - name: "TenantId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Tenant"
...
definitions:
...
  ApiResponse:
    type: "object"
    properties:
      RequestId:
        type: "string"
        format: "int32"
      Message:
        type: "string"
        example: "everything is ok"
    example:
      RequestId: "RequestId"
      Message: "everything is ok"
...
  tenantAddRequest:
    type: "object"
    properties:
      TenantId:
        type: "string"
      Password:
        type: "string"
...
  tenantResponse:
    type: "object"
    properties:
      requestId:
        type: "string"
        example: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
      result:
        type: "object"
        properties: {}
    example:
      result: "{}"
      requestId: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
...